+++ src/components/containers/results-list.js
@@ -6,47 +6,6 @@ import TransparentButton from 'Components/buttons/transparent-button';
 import Button from 'Components/buttons/button';
 import styles from './results-list.styl';
 
-export const ScrollResult = ({ active, children }) => {
-  const ref = useRef();
-  useEffect(() => {
-    if (active && ref.current.scrollIntoView) {
-      ref.current.scrollIntoView({ behavior: 'smooth', block: 'nearest' });
-    }
-  }, [active]);
-  return <div ref={ref}>{children}</div>;
-};
-
-export function useActiveIndex(items, onSelect) {
-  const [activeIndex, setActiveIndex] = useState(-1);
-  useEffect(() => {
-    setActiveIndex(-1);
-  }, [items]);
-  const onKeyDown = (e) => {
-    if (e.key === 'ArrowUp') {
-      e.preventDefault();
-      setActiveIndex((prev) => {
-        if (prev < 0) {
-          return items.length - 1;
-        }
-        return prev - 1;
-      });
-    } else if (e.key === 'ArrowDown') {
-      e.preventDefault();
-      setActiveIndex((prev) => {
-        if (prev === items.length - 1) {
-          return -1;
-        }
-        return prev + 1;
-      });
-    } else if (e.key === 'Enter') {
-      e.preventDefault();
-      if (items[activeIndex]) {
-        onSelect(items[activeIndex]);
-      }
-    }
-  };
-  return { activeIndex, onKeyDown };
-}
 
 const ResultsList = ({ scroll, items, className, children }) => (
   <div className={classnames(scroll && styles.scrollContainer, className)}>
